feat(#6) model 분리
